+++ sdk/ts/src/index.ts
@@ -1,3 +1,4 @@
 export * as constants from "./constants.js";
 export * as pda from "./pda.js";
 export * from "./types.js";
+export * from "./platform.js";